Removing CMake ZERO_CHECK project
Removing automatically generated project
that was preventing the out-of-the-box
build from completing successfully.

Building the main project will work now,
although a build on BUIL_ALL project
will continue to fail.

What was not possible to accomplish was to
define wihtin the CMakeList.txt to
add the following entry for the ZERO_CHECK
project:
<TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>